United hsv palette listeners
