+++ include/palette_view.h
@@ -5,8 +5,7 @@
 
 #include "include/glview.h"
 #include "include/layout.h"
-#include "include/value_palette_listener.h"
-#include "include/hue_sat_palette_listener.h"
+#include "include/hsv_palette_listener.h"
 
 class PaletteView : public GLView {
  public:
@@ -57,8 +56,8 @@ class PaletteView : public GLView {
   void UpdateVPaletteColors();
 
   Layout layout_;
-  ValuePaletteListener v_palette_listener_;
-  HueSatPaletteListener hs_palette_listener_;
+  HSVPaletteListener v_palette_listener_;
+  HSVPaletteListener hs_palette_listener_;
 
   int* hs_palette_points_;
   unsigned n_hs_palette_points_;
